use code snippets for retrieve & query how-tos
