+++ code-snippets/how-to/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "web3-storage-how-to-store-example",
+  "name": "web3-storage-how-to-examples",
   "version": "1.0.0",
-  "description": "Code snippets for https://docs.web3.storage/how-to/store",
+  "description": "Code snippets for https://docs.web3.storage how-to articles",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
