+++ code-snippets/how-to/platform-browser.js
