+++ docs/how-tos/store.md
@@ -30,21 +30,9 @@ npm install web3.storage
 
 ## Creating a client instance
 
-First import the `Web3Storage` constructor from the `web3.storage` package:
+First we need to create a `Web3.Storage` client object, passing in an API token to its [constructor][reference-js-constructor]:
 
-<<<@/code-snippets/how-to/store/index.js#import-client
-
-To create a `Web3Storage` client object, we need to pass an access token into the [constructor][reference-js-constructor].
-
-Generally speaking, it's best to read credentials like access tokens from an environment variable or configuration file that's kept outside of source control. 
-
-Here's an example of a helper function that will read in a token from the `WEB3STORAGE_TOKEN` environment variable:
-
-<<<@/code-snippets/how-to/store/index.js#getAccessToken
-
-Now it's easy to create a new client object:
-
-<<<@/code-snippets/how-to/store/index.js#makeStorageClient
+<<<@/code-snippets/how-to/index.js#makeStorageClient
 
 ## Preparing files for upload
 
@@ -58,11 +46,11 @@ There are a few different ways of creating `File` objects available, depending o
 
 In the browser, you can use a [file input element][mdn-file-input] to allow the user to select files for upload:
 
-<<<@/code-snippets/how-to/store/platform-browser.js#getFiles
+<<<@/code-snippets/how-to/platform-browser.js#getFiles
 
 You can also manually create `File` objects using the native `File` constructor provided by the browser runtime:
 
-<<<@/code-snippets/how-to/store/platform-browser.js#makeFileObjects
+<<<@/code-snippets/how-to/platform-browser.js#makeFileObjects
 
 :::
 
@@ -70,13 +58,13 @@ You can also manually create `File` objects using the native `File` constructor 
 
 In Node.js, the `web3.storage` package exports some helpful utility functions from the [`files-from-path` module](https://www.npmjs.com/package/files-from-path) that allow you to easily read `File` objects from the local file system. The `getFilesFromPath` helper asynchronously returns an array of `File`s that you can use directly with the `put` client method:
 
-<<<@/code-snippets/how-to/store/platform-node.js#getFiles
+<<<@/code-snippets/how-to/platform-node.js#getFiles
 
 If you expect to be loading a lot of large files, you may be better served by the [`filesFromPath` helper](https://github.com/web3-storage/files-from-path#filesfrompath). It reduces memory pressure by `yield`ing `File` objects one by one as they're loaded from disk, instead of loading everything into memory. You can then issue multiple `put` requests to send each file to Web3.Storage.
 
 You can also manually create `File` objects by importing a Node.js implementation of `File` from the `web3.storage` package:
 
-<<<@/code-snippets/how-to/store/platform-node.js#makeFileObjects
+<<<@/code-snippets/how-to/platform-node.js#makeFileObjects
 
 :::
 
@@ -91,7 +79,7 @@ You can also manually create `File` objects by importing a Node.js implementatio
 
 Once you have an array of `File`s, uploading is simple:
 
-<<<@/code-snippets/how-to/store/index.js#storeFiles
+<<<@/code-snippets/how-to/index.js#storeFiles
 
 ::: warning IMPORTANT
 Deleting files from the Web3.Storage site's [Files page][site-files] will remove them from the file listing for your account, but that doesn't prevent nodes on the [decentralized storage network][concepts-decentralized-storage] from retaining copies of the data indefinitely. Do not use Web3.Storage for data that may need to be permanently deleted in the future.
